pyproject.toml
[tool.poetry]
name = "disraptors"
version = "0.1.0"
description = ""
authors = ["qhongh <qianhong.huang.12@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
langchain = "^0.1.16"
grandalf = "^0.8"
matplotlib = "^3.8.4"
langchainhub = "^0.1.15"
praw = "^7.7.1"
beautifulsoup4 = "^4.12.3"
sentence-transformers = "^2.6.1"
pinecone-client = "^3.2.2"
pinecone-text = "^0.9.0"
langchain-openai = "^0.1.3"
langchain-text-splitters = "^0.0.1"
langchain-pinecone = "^0.1.0"
jupyterlab = "^4.1.6"
langchain-anthropic = "^0.1.8"
llama-index = "^0.10.29"
llama-index-embeddings-huggingface = "^0.2.0"
llama-index-embeddings-openai = "^0.1.7"
llama-index-llms-openai = "^0.1.15"
llama-index-llms-langchain = "^0.1.3"
llama-index-embeddings-langchain = "^0.1.2"
langchain-chroma = "^0.1.0"
bertopic = "^0.16.1"
nest-asyncio = "^1.6.0"
llama-index-retrievers-bm25 = "^0.1.3"
st-pages = "^0.4.5"
streamlit = "^1.33.0"
watchdog = "^4.0.0"
openpyxl = "^3.1.2"

[tool.poetry.dev-dependencies]
black = "^24.4.0"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
